Updated OpenGL and PortAudio thread synchronization in Cocoa

+++ macosx/DocumentController.h
@@ -29,7 +29,7 @@
 	
 	void *portAudioHAL;
 	
-	int disableMenuBarCount;
+	NSInteger disableMenuBarCount;
 }
 
 - (NSArray *)diskImagePathExtensions;
